src/requirements.txt
aiofiles==23.2.1
celery==5.3.6
ex4nicegui==0.4.4
fastapi==0.110.0
httpx==0.27.0
loguru==0.7.2
nicegui==1.4.17
passlib==1.7.4
Pillow==10.1.0
Pillow==10.2.0
pipreqs==0.4.13
pydantic==2.6.3
pydantic_settings==2.2.1
python_jose==3.3.0
pytz==2023.3.post1
redis==4.6.0
SQLAlchemy==1.4.51
starlette==0.37.2
uvicorn==0.27.1
websockets==11.0.3
